requirements.txt
numpy>=1.23
opencv-python>=4.8
matplotlib>=3.7
scikit-image>=0.21
pywavelets>=1.4
pillow>=10.0
scipy>=1.10
click>=8.1